pyproject.toml
[tool.poetry]
name = "Prophet-Forecasting-For-Portfolio-Optimisation"
version = "0.1.0"
description = "Prophet-based stock price forecasting with Markowitz portfolio optimisation"
authors = ["egor@egorhowell.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.0"
pandas = "^2.0.0"
scipy = "^1.11.0"
yfinance = "^0.2.28"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
prophet = "^1.1.5"
supabase = "^2.23.0"
streamlit = "^1.28.0"
pandas-market-calendars = "^5.1.3"
plotly = "^6.4.0"
altair = "^5.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.0"
black = "^23.7.0"
ruff = "^0.0.285"
mypy = "^1.5.0"
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
]


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
]

